Fix segfault

+++ source/qcommon/utf8.cpp
@@ -59,7 +59,7 @@ const char * StrChrUTF8( const char * str, uint32_t needle ) {
 	uint32_t state = 0;
 	for( const char * p = str; *p != '\0'; p++ ) {
 		uint32_t c;
-		if( DecodeUTF8( &state, &c, *p ) != 0 )
+		if( DecodeUTF8( &state, &c, uint8_t( *p ) ) != 0 )
 			continue;
 		if( c == needle )
 			return p;
